monitor_bot/requirements.txt
python-telegram-bot[job-queue]
aiohttp
aiosqlite
apscheduler
python-dotenv
